--- eventmesh-connector-plugin/eventmesh-connector-mongodb/src/test/resources/mongodb-client.properties
@@ -1,29 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-####################### mongodb server ##################
-eventMesh.server.mongodb.connector.type=STANDALONE
-##like: mongodb://root:root@127.0.0.1:27018,127.0.0.1:27019
-eventMesh.server.mongodb.url=mongodb://127.0.0.1:27018
-eventMesh.server.mongodb.database=db
-eventMesh.server.mongodb.collection=col
-
-#eventMesh.server.mongodb.connector.type=REPLICA_SET
-##like: mongodb://root:root@127.0.0.1:27018,127.0.0.1:27019
-#eventMesh.server.mongodb.url=mongodb://root:root@127.0.0.1:27018,127.0.0.1:27019
-#eventMesh.server.mongodb.database=db
-#eventMesh.server.mongodb.collection=col
