+++ eventmesh-storage/eventmesh-storage-mongodb/src/test/java/org/apache/eventmesh/storage/mongodb/MongodbServerTest.java
@@ -1,0 +1,71 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package org.apache.eventmesh.storage.mongodb;
+
+import org.bson.Document;
+import org.junit.After;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+
+import com.mongodb.client.MongoClient;
+import com.mongodb.client.MongoClients;
+import com.mongodb.client.MongoCollection;
+
+import de.bwaldvogel.mongo.MongoServer;
+import de.bwaldvogel.mongo.backend.memory.MemoryBackend;
+
+public class MongodbServerTest {
+    private MongoCollection<Document> collection;
+    private MongoClient client;
+    private MongoServer server;
+
+    @Before
+    public void setUp() {
+        server = new MongoServer(new MemoryBackend());
+
+        // optionally:
+        // server.enableSsl(key, keyPassword, certificate);
+        // server.enableOplog();
+
+        // bind on a random local port
+        server.bind("127.0.0.1", 27011);
+        String connectionString = server.getConnectionString();
+
+        client = MongoClients.create(connectionString);
+        collection = client.getDatabase("testdb").getCollection("testcollection");
+    }
+
+    @After
+    public void tearDown() {
+        client.close();
+        server.shutdown();
+    }
+
+    @Test
+    public void testSimpleInsertQuery() {
+        Assert.assertEquals(collection.countDocuments(), 0);
+
+        // creates the database and collection in memory and insert the object
+        Document obj = new Document("_id", 1).append("key", "value");
+        collection.insertOne(obj);
+
+        Assert.assertEquals(collection.countDocuments(), 1L);
+        Assert.assertEquals(collection.find().first(), obj);
+    }
+}
