+++ settings.gradle
@@ -26,6 +26,7 @@ include 'eventmesh-spi'
 include 'eventmesh-storage:eventmesh-storage-api'
 include 'eventmesh-storage:eventmesh-storage-standalone'
 include 'eventmesh-storage:eventmesh-storage-kafka'
+include 'eventmesh-storage:eventmesh-storage-mongodb'
 include 'eventmesh-storage:eventmesh-storage-pulsar'
 include 'eventmesh-storage:eventmesh-storage-pravega'
 include 'eventmesh-storage:eventmesh-storage-redis'
@@ -33,7 +34,6 @@ include 'eventmesh-storage:eventmesh-storage-rocketmq'
 include 'eventmesh-storage:eventmesh-storage-rabbitmq'
 include 'eventmesh-connector-plugin:eventmesh-connector-api'
 include 'eventmesh-connector-plugin:eventmesh-connector-knative'
-include 'eventmesh-connector-plugin:eventmesh-connector-mongodb'
 include 'eventmesh-connector-plugin:eventmesh-connector-pravega'
 include 'eventmesh-connector-plugin:eventmesh-connector-rocketmq'
 include 'eventmesh-connector-plugin:eventmesh-connector-rabbitmq'
